runner/requirements.txt
# Python version requirement
python>=3.12.0

# Dependencies
python-dotenv>=1.0.0 